modified testFiles and code review feedback
