--- src/__test__/helper/types/TodoResponseType.ts
@@ -1,7 +1,0 @@
-export type TodoResponseType = {
-  id: number;
-  title: string;
-  body: string;
-  createdAt: Date;
-  updateAt: Date;
-};
